check disk persist before lru delete

+++ src/main/java/com/lru/memory/disk/cache/LRUCache.java
@@ -26,7 +26,7 @@ public class LRUCache<K extends String, V> extends LinkedHashMap<K, V> {
     @Override
     protected boolean removeEldestEntry(Entry<K, V> eldest) {        
         if(this.size() > this.cachesize){
-            deleteFromFilesystem(eldest);
+            if(this.dirLocate.isDiskPersistent()) deleteFromFilesystem(eldest);
             return true;
         }        
         return false;
